Mock db used for testing.
