+++ templates/express_mysql/index.js
@@ -1,35 +1,19 @@
-import "dotenv/config.js";
-
-import express, { json } from "express";
-import helmet from "helmet";
 import { appendFileSync } from "fs";
-import cors from "cors";
 
+import makeApp from "./app.js"
+import database from "./db/database.js"
 import { initLog } from "./logs/initLog.js";
-import sampleRouter from "./router/sampleRouter.js";
+import connectToDb from "./connection/normalConnection.js"
+import reInitDb from "./db/reInitDb.js"
 import { appConfig } from "./config/appConfig.js";
-import connectToDb from "./connection/normalConnection.js";
-import reInitDb from "./db/reInitDb.js";
 
-const app = express();
-
-// Helmet sets HTTP headers for security.
-app.use(helmet());
-
-app.use(cors());
-app.use(json());
-
-// Disable the X-Powered-By header to make it harder
-// for attackers to find the tech stack.
-app.disable("x-powered-by");
-
-app.use(appConfig.router.SAMPLE_PREFIX, sampleRouter);
+const app = makeApp(database);
 
 initLog();
 const db = connectToDb();
 await reInitDb(db);
 
-const server = app.listen(appConfig.PORT, (err) => {
+app.listen(appConfig.PORT, (err) => {
     if (err) {
         const timeStamp = new Date().toLocaleString();
         const errMessage = `[ERROR]: ${timeStamp} - ${err.message}`;
@@ -44,5 +28,3 @@ const server = app.listen(appConfig.PORT, (err) => {
         );
     }
 });
-
-export { app, server };
